+++ src/devices/src/virtio/block/device.rs
@@ -19,7 +19,7 @@ use logger::{error, warn, IncMetric, METRICS};
 use rate_limiter::{BucketUpdate, RateLimiter, TokenType};
 use utils::eventfd::EventFd;
 use virtio_gen::virtio_blk::*;
-use vm_memory::{Bytes, GuestMemoryError, GuestMemoryMmap};
+use vm_memory::{Bytes, GuestMemoryMmap};
 
 use super::{
     super::{ActivateResult, DeviceState, Queue, VirtioDevice, TYPE_BLOCK, VIRTIO_MMIO_INT_VRING},
@@ -283,8 +283,7 @@ impl Block {
         let queue = &mut self.queues[queue_index];
         let mut used_any = false;
         while let Some(head) = queue.pop(mem) {
-            let len;
-            match Request::parse(&head, mem) {
+            let len = match Request::parse(&head, mem) {
                 Ok(request) => {
                     // If limiter.consume() fails it means there is no more TokenType::Ops
                     // budget and rate limiting is in effect.
@@ -315,59 +314,29 @@ impl Block {
                         }
                     }
 
-                    let status = match request.execute(&mut self.disk, mem) {
-                        Ok(l) => {
-                            // Account for the status byte as well.
-                            // With a non-faulty driver, we shouldn't get to the point where we
-                            // overflow here (since data len must be a multiple of 512 bytes, so
-                            // it can't be u32::MAX). In the future, this should be fixed at the
-                            // request parsing level, so no data will actually be transferred in
-                            // scenarios like this one.
-                            if let Some(l) = l.checked_add(1) {
-                                len = l;
-                                VIRTIO_BLK_S_OK
-                            } else {
-                                len = l;
-                                VIRTIO_BLK_S_IOERR
-                            }
-                        }
-                        Err(e) => {
-                            METRICS.block.invalid_reqs_count.inc();
-                            match e {
-                                ExecuteError::Read(GuestMemoryError::PartialBuffer {
-                                    completed,
-                                    expected,
-                                }) => {
-                                    error!(
-                                        "Failed to execute virtio block read request: can only \
-                                        write {} of {} bytes.",
-                                        completed, expected
-                                    );
-                                    METRICS.block.read_bytes.add(completed);
-                                    // This can not overflow since `completed` < data len which is
-                                    // an u32.
-                                    len = completed as u32 + 1;
-                                }
-                                _ => {
-                                    error!("Failed to execute virtio block request: {:?}", e);
-                                    // Status byte only.
-                                    len = 1;
-                                }
-                            };
-                            e.status()
-                        }
-                    };
+                    let status = Status::from_result(request.execute(&mut self.disk, mem));
+                    let virtio_blk_status = status.virtio_blk_status();
+                    let num_used_bytes = status.num_used_bytes();
+                    if let Status::Err(err_status) = status {
+                        METRICS.block.invalid_reqs_count.inc();
+                        error!(
+                            "Failed to execute {:?} virtio block request: {:?}",
+                            request.request_type, err_status
+                        );
+                    }
 
-                    if let Err(e) = mem.write_obj(status, request.status_addr) {
+                    if let Err(e) = mem.write_obj(virtio_blk_status, request.status_addr) {
                         error!("Failed to write virtio block status: {:?}", e)
                     }
+
+                    num_used_bytes
                 }
                 Err(e) => {
                     error!("Failed to parse available descriptor chain: {:?}", e);
                     METRICS.block.execute_fails.inc();
-                    len = 0;
+                    0
                 }
-            }
+            };
 
             queue.add_used(mem, head.index, len).unwrap_or_else(|e| {
                 error!(
@@ -527,6 +496,8 @@ impl VirtioDevice for Block {
 #[cfg(test)]
 pub(crate) mod tests {
     use std::fs::metadata;
+    use std::io::Read;
+    use std::os::unix::ffi::OsStrExt;
     use std::thread;
     use std::time::Duration;
     use std::u32;
@@ -835,14 +806,46 @@ pub(crate) mod tests {
         let data_addr = GuestAddress(vq.dtable[1].addr.get());
         let status_addr = GuestAddress(vq.dtable[2].addr.get());
 
-        // Write.
+        let empty_data = vec![0; 512];
+        let rand_data = utils::rand::rand_alphanumerics(1024).as_bytes().to_vec();
+
+        // Write with invalid data len (not a multiple of 512).
         {
             mem.write_obj::<u32>(VIRTIO_BLK_T_OUT, request_type_addr)
                 .unwrap();
-            // Make data read only, 8 bytes in len, and set the actual value to be written.
+            // Make data read only, 512 bytes in len, and set the actual value to be written.
             vq.dtable[1].flags.set(VIRTQ_DESC_F_NEXT);
-            vq.dtable[1].len.set(8);
-            mem.write_obj::<u64>(123_456_789, data_addr).unwrap();
+            vq.dtable[1].len.set(511);
+            mem.write_slice(&rand_data[..511], data_addr).unwrap();
+
+            invoke_handler_for_queue_event(&mut block);
+
+            assert_eq!(vq.used.idx.get(), 1);
+            assert_eq!(vq.used.ring[0].get().id, 0);
+            assert_eq!(vq.used.ring[0].get().len, 1);
+            assert_eq!(
+                mem.read_obj::<u32>(status_addr).unwrap(),
+                VIRTIO_BLK_S_IOERR
+            );
+
+            // Check that the data wasn't written to the file
+            let mut buf = [0u8; 512];
+            block.disk.file.seek(SeekFrom::Start(0)).unwrap();
+            block.disk.file.read_exact(&mut buf).unwrap();
+            assert_eq!(buf, empty_data.as_slice());
+        }
+
+        // Write.
+        {
+            vq.used.idx.set(0);
+            set_queue(&mut block, 0, vq.create_queue());
+
+            mem.write_obj::<u32>(VIRTIO_BLK_T_OUT, request_type_addr)
+                .unwrap();
+            // Make data read only, 512 bytes in len, and set the actual value to be written.
+            vq.dtable[1].flags.set(VIRTQ_DESC_F_NEXT);
+            vq.dtable[1].len.set(512);
+            mem.write_slice(&rand_data[..512], data_addr).unwrap();
 
             check_metric_after_block!(
                 &METRICS.block.write_count,
@@ -856,6 +859,37 @@ pub(crate) mod tests {
             assert_eq!(mem.read_obj::<u32>(status_addr).unwrap(), VIRTIO_BLK_S_OK);
         }
 
+        // Read with invalid data len (not a multiple of 512).
+        {
+            vq.used.idx.set(0);
+            set_queue(&mut block, 0, vq.create_queue());
+
+            mem.write_obj::<u32>(VIRTIO_BLK_T_IN, request_type_addr)
+                .unwrap();
+            vq.dtable[1]
+                .flags
+                .set(VIRTQ_DESC_F_NEXT | VIRTQ_DESC_F_WRITE);
+            vq.dtable[1].len.set(511);
+            mem.write_slice(empty_data.as_slice(), data_addr).unwrap();
+
+            invoke_handler_for_queue_event(&mut block);
+
+            assert_eq!(vq.used.idx.get(), 1);
+            assert_eq!(vq.used.ring[0].get().id, 0);
+
+            // Only status byte length.
+            assert_eq!(vq.used.ring[0].get().len, 1);
+            assert_eq!(
+                mem.read_obj::<u32>(status_addr).unwrap(),
+                VIRTIO_BLK_S_IOERR
+            );
+
+            // Check that no data was read.
+            let mut buf = [0u8; 512];
+            mem.read_slice(&mut buf, data_addr).unwrap();
+            assert_eq!(buf, empty_data.as_slice());
+        }
+
         // Read.
         {
             vq.used.idx.set(0);
@@ -866,6 +900,8 @@ pub(crate) mod tests {
             vq.dtable[1]
                 .flags
                 .set(VIRTQ_DESC_F_NEXT | VIRTQ_DESC_F_WRITE);
+            vq.dtable[1].len.set(512);
+            mem.write_slice(empty_data.as_slice(), data_addr).unwrap();
 
             check_metric_after_block!(
                 &METRICS.block.read_count,
@@ -878,6 +914,11 @@ pub(crate) mod tests {
             // Added status byte length.
             assert_eq!(vq.used.ring[0].get().len, vq.dtable[1].len.get() + 1);
             assert_eq!(mem.read_obj::<u32>(status_addr).unwrap(), VIRTIO_BLK_S_OK);
+
+            // Check that the data is the same that we wrote before
+            let mut buf = [0u8; 512];
+            mem.read_slice(&mut buf, data_addr).unwrap();
+            assert_eq!(buf, &rand_data[..512]);
         }
 
         // Read with error.
@@ -890,6 +931,7 @@ pub(crate) mod tests {
             vq.dtable[1]
                 .flags
                 .set(VIRTQ_DESC_F_NEXT | VIRTQ_DESC_F_WRITE);
+            mem.write_slice(empty_data.as_slice(), data_addr).unwrap();
 
             let size = block.disk.file.seek(SeekFrom::End(0)).unwrap();
             block.disk.file.set_len(size / 2).unwrap();
@@ -907,7 +949,11 @@ pub(crate) mod tests {
                 mem.read_obj::<u32>(status_addr).unwrap(),
                 VIRTIO_BLK_S_IOERR
             );
-            assert_eq!(mem.read_obj::<u64>(data_addr).unwrap(), 123_456_789);
+
+            // Check that no data was read.
+            let mut buf = [0u8; 512];
+            mem.read_slice(&mut buf, data_addr).unwrap();
+            assert_eq!(buf, empty_data.as_slice());
         }
 
         // Partial buffer error on read.
@@ -939,7 +985,11 @@ pub(crate) mod tests {
                 mem.read_obj::<u32>(status_addr).unwrap(),
                 VIRTIO_BLK_S_IOERR
             );
-            assert_eq!(mem.read_obj::<u64>(data_addr).unwrap(), 123_456_789);
+
+            // Check that no data was read since we can't read past the end of the file.
+            let mut buf = [0u8; 512];
+            mem.read_slice(&mut buf, data_addr).unwrap();
+            assert_eq!(buf, empty_data.as_slice());
         }
 
         {
@@ -956,6 +1006,9 @@ pub(crate) mod tests {
             mem.write_obj(1, GuestAddress(request_type_addr.0 + 8))
                 .unwrap();
 
+            block.disk.file.seek(SeekFrom::Start(512)).unwrap();
+            block.disk.file.write_all(&rand_data[512..]).unwrap();
+
             invoke_handler_for_queue_event(&mut block);
 
             assert_eq!(vq.used.idx.get(), 1);
@@ -968,6 +1021,11 @@ pub(crate) mod tests {
                 mem.read_obj::<u32>(status_addr).unwrap(),
                 VIRTIO_BLK_S_IOERR
             );
+
+            // Check that we correctly read the second file sector.
+            let mut buf = [0u8; 512];
+            mem.read_slice(&mut buf, data_addr).unwrap();
+            assert_eq!(buf, rand_data[512..]);
         }
     }
 
@@ -1094,18 +1152,18 @@ pub(crate) mod tests {
         let data_addr = GuestAddress(vq.dtable[1].addr.get());
         let status_addr = GuestAddress(vq.dtable[2].addr.get());
 
-        // Create bandwidth rate limiter that allows only 80 bytes/s with bucket size of 8 bytes.
-        let mut rl = RateLimiter::new(8, 0, 100, 0, 0, 0).unwrap();
+        // Create bandwidth rate limiter that allows only 5120 bytes/s with bucket size of 8 bytes.
+        let mut rl = RateLimiter::new(512, 0, 100, 0, 0, 0).unwrap();
         // Use up the budget.
-        assert!(rl.consume(8, TokenType::Bytes));
+        assert!(rl.consume(512, TokenType::Bytes));
 
         set_rate_limiter(&mut block, rl);
 
         mem.write_obj::<u32>(VIRTIO_BLK_T_OUT, request_type_addr)
             .unwrap();
-        // Make data read only, 8 bytes in len, and set the actual value to be written
+        // Make data read only, 512 bytes in len, and set the actual value to be written
         vq.dtable[1].flags.set(VIRTQ_DESC_F_NEXT);
-        vq.dtable[1].len.set(8);
+        vq.dtable[1].len.set(512);
         mem.write_obj::<u64>(123_456_789, data_addr).unwrap();
 
         // Following write procedure should fail because of bandwidth rate limiting.
@@ -1173,9 +1231,9 @@ pub(crate) mod tests {
 
         mem.write_obj::<u32>(VIRTIO_BLK_T_OUT, request_type_addr)
             .unwrap();
-        // Make data read only, 8 bytes in len, and set the actual value to be written.
+        // Make data read only, 512 bytes in len, and set the actual value to be written.
         vq.dtable[1].flags.set(VIRTQ_DESC_F_NEXT);
-        vq.dtable[1].len.set(8);
+        vq.dtable[1].len.set(512);
         mem.write_obj::<u64>(123_456_789, data_addr).unwrap();
 
         // Following write procedure should fail because of ops rate limiting.